refactor(inferno-test-utils): Adjust imports and types
