+++ packages/inferno-create-class/package.json
