+++ packages/inferno-devtools/package.json
