+++ packages/inferno-hyperscript/package.json
