+++ packages/inferno-mobx/package.json
