+++ packages/inferno-redux/package.json
