+++ packages/inferno-test-utils/inferno-test-utils.js
@@ -1,2 +1,2 @@
-module.exports = require('inferno/dist/inferno-test-utils.node');
+module.exports = require('./dist/inferno-test-utils.node');
 module.exports.default = module.exports;
